Fixed line deletion + code refactor
